m1/N1: the wire's quantiser casts FloorDiv's int64 back to int32
Neuron::FloorDiv takes and returns int64_t, and returning it from an int32_t
function is C4244 - a warning MSVC turns into an error under /W4 /WX, which run
91 did. The native harness this was written against compiles with -w for speed,
which is exactly how it got past a local run; it now has a warnings-on syntax
pass beside it.

+++ Net/Records.h
@@ -54,7 +54,7 @@ inline constexpr std::int32_t SUBUNITS_PER_WIRE_UNIT = Neuron::SUBUNITS_PER_WORL
 /// +1 subunit must not both become wire 0 while every other step is 64 wide.
 [[nodiscard]] constexpr std::int32_t WireFromSubunits(std::int32_t _subunits) noexcept
 {
-  return Neuron::FloorDiv(_subunits, SUBUNITS_PER_WIRE_UNIT);
+  return static_cast<std::int32_t>(Neuron::FloorDiv(_subunits, SUBUNITS_PER_WIRE_UNIT));
 }
 
 /// The middle of the quarter-unit the wire named, so that the round trip's error is half a step
